Update Cube.py
Creates a numpy 3x3x3 array. It works and creates the array but when trying to access the objects data inside of the array in does not work. Once again might be easier to switch to Java. Is it going to be fun? No. But I also have to search up everything on google anyways while making these arrays and I've seen several times that it is a bad idea to use objects inside of numpy whereas that is what Java is meant for so might be easier to switch to that instead of having a pain in the ass time the whole project. If you can somehow access the data of objects inside the numpy array then we should be set and only have to iterate over the array and give the cubelets information before being able to manipulate the cube. Once again, Java.